# Optimization: launch fewer threads and loop over nv in init_qfrc_constraint