--- a/mujoco_warp/_src/solver.py
+++ b/mujoco_warp/_src/solver.py
@@ -1757,6 +1757,37 @@ def update_constraint_init_qfrc_constraint(
     efc_J_in[efcid, dofid] * efc_force_in[efcid],
   )
 
+@wp.kernel
+def update_constraint_init_qfrc_constraintv2(
+  # Model in:
+  nv: int,
+  # Data in:
+  nefc_in: wp.array(dtype=int),
+  efc_worldid_in: wp.array(dtype=int),
+  efc_J_in: wp.array2d(dtype=float),
+  efc_force_in: wp.array(dtype=float),
+  efc_done_in: wp.array(dtype=bool),
+  # Data out:
+  qfrc_constraint_out: wp.array2d(dtype=float),
+):
+  efcid = wp.tid()
+
+  if efcid >= nefc_in[0]:
+    return
+
+  worldid = efc_worldid_in[efcid]
+
+  if efc_done_in[worldid]:
+    return
+
+  force = efc_force_in[efcid]
+  for i in range(nv):
+    wp.atomic_add(
+      qfrc_constraint_out[worldid],
+      i,
+      efc_J_in[efcid, i] * force,
+    )
+
 
 @wp.kernel
 def update_constraint_gauss_cost(
@@ -1893,10 +1924,17 @@ def _update_constraint(m: types.Model, d: types.Data):
     outputs=[d.qfrc_constraint],
   )
 
+  #wp.launch(
+  #  update_constraint_init_qfrc_constraint,
+  #  dim=(m.nv, d.njmax),
+  #  inputs=[d.nefc, d.efc.worldid, d.efc.J, d.efc.force, d.efc.done],
+  #  outputs=[d.qfrc_constraint],
+  #)
+
   wp.launch(
-    update_constraint_init_qfrc_constraint,
-    dim=(m.nv, d.njmax),
-    inputs=[d.nefc, d.efc.worldid, d.efc.J, d.efc.force, d.efc.done],
+    update_constraint_init_qfrc_constraintv2,
+    dim=(d.njmax),
+    inputs=[m.nv,d.nefc, d.efc.worldid, d.efc.J, d.efc.force, d.efc.done],
     outputs=[d.qfrc_constraint],
   )