Seems to have a good effect on many assets, especially if you look at low-level traces. Makes sense if you think about it - we only need to load the force once.

Total time spent in this kernel for some assets (main -> this):
humanoid: 4.4ms -> 1.9ms
n_humanoids: 333ms -> 29.8ms
internal asset with nv==70: 94ms -> 14.58ms
Ant: 2.7s -> 294ms
cartpole: 693ms -> 286ms